+++ app/models.py
@@ -161,9 +161,6 @@ db.create_all()
 
 
 
-def get_user_submissions_for_assignment(user_id, assignment_id):
-    return Submission.query.filter_by(userId=user_id, assignmentId=assignment_id).all()
-
 
 
 def register(name, email, student_number, password,role):
@@ -226,93 +223,6 @@ def register(name, email, student_number, password,role):
 
     return True
 
-def enrollInCourse(course_code, enrollment_password, user):
-    if user.role != 1:
-        return False
-
-    course = Course.query.filter_by(courseCode=course_code).first()
-        
-    if course and course.enrollmentPassword == enrollment_password:
-        # Assume user_id is the ID of the currently logged-in student
-        user_course = UserCourse(courseId=course.id, userId=user.id, userRole=1)
-        db.session.add(user_course)
-        db.session.commit()
-        return True
-    return False
-
-def enrollTaInCourse(courseId, user, ta_email):
-    if user.role != 3:
-        return False
-    
-    course = Course.query.filter(Course.id == courseId).first()
-
-    user = User.query.filter(User.email == ta_email).first()
-    if course and user and user.role == 2:
-        user_course = UserCourse(courseId=course.id, userId=user.id, userRole=2)
-        db.session.add(user_course)
-        db.session.commit()
-        return True
-
-    return False
-
-
-
-def findUserById(user_id):
-    user = User.query.filter(User.id == user_id).first()
-    return user
-
-
-def getCourseById(courseId, user):
-    #find course
-    course = Course.query.filter(Course.id == courseId).all()
-    if len(course) < 1:
-        return None
-    
-    #check that the user can view this course
-    if not checkUserInCourse:
-        return None
-    
-    return course[0]
-
-def getAssignmentsById(assignmentId, user):
-    #find assignment
-    assignment = Assignment.query.filter(Assignment.id == assignmentId).all()
-    if len(assignment) < 1:
-        return None
-
-    return assignment[0]
-
-def getUsersForCourse(courseId, user):
-    try:
-        # Explicitly specify the join condition and filter by userId
-        users = db.session.query(User).\
-            join(UserCourse, UserCourse.userId == User.id).\
-            filter(UserCourse.courseId == courseId).all()
-        return users
-    except Exception as e:
-        print(f"Error fetching users for course {courseId}: {e}")
-        return []
-
-def getAssignmentsForCourse(courseId, user):
-    try:
-        # Explicitly specify the join condition and filter by userId
-        assignments = db.session.query(Assignment).\
-            join(CourseAssignment, CourseAssignment.assignmentId == Assignment.id).\
-            filter(CourseAssignment.courseId == courseId).all()
-        return assignments
-    except Exception as e:
-        print(f"Error fetching assignments for user {user.id}: {e}")
-        return []
-
-def find_assignments(user):
-    assignments = Assignment.query.filter(Assignment.createdBy == user.id).all()
-    return assignments
-
-def checkUserInCourse(courseId, userId):
-    userCourse = UserCourse.query.filter(and_(UserCourse.courseId == courseId, UserCourse.userId == userId)).all()
-    
-    return len(userCourse) > 0
-    
 
 def login(email, password):
     '''
@@ -336,350 +246,8 @@ def login(email, password):
         return valids[0]
     else:
         return False
-    
-def create_course(name, course_code, year, semester, start_date, end_date, userId):
-    # Generate a unique password for course enrollment
-    enrollment_password = token_urlsafe(16)  # Generates a URL-safe text string, containing 16 random bytes
-    
-    # Check if the generated password is truly unique
-    while Course.query.filter_by(enrollmentPassword=enrollment_password).first() is not None:
-        # If the generated password is not unique, generate a new one
-        enrollment_password = token_urlsafe(16)
-
-    course = Course(name=name, courseCode=course_code, year=year,
-                    semester=semester, startDate=start_date, endDate=end_date, 
-                    createdBy=userId, enrollmentPassword=enrollment_password) # Add the password here
-    db.session.add(course)
-    db.session.commit()
-
-    userCourse = UserCourse(courseId=course.id, userId=userId, userRole=3)
-    db.session.add(userCourse)
-    db.session.commit()
-    return True
-
-def addSubmissionLog(filename_with_user_id, user, assignment_id):
-    pastSubmissions = Submission.query.filter(and_( Submission.userId == user.id, Submission.assignmentId == assignment_id)).all()
-    for pastSubmission in pastSubmissions:
-        #overwrite each past submission to only store the latest
-        pastSubmission.overwritten = True
-    submission = Submission(assignmentId = assignment_id, userId=user.id, fileName=filename_with_user_id, submissionDate=func.now())
-    db.session.add(submission)
-    db.session.commit()
-    db.session.flush() 
-    return submission
-
-def getSubmissions(course_id, user_id):
-    submissions = Submission.query.filter(Submission.userId == user_id)\
-                                  .join(Assignment, Assignment.id == Submission.assignmentId)\
-                                  .join(CourseAssignment, CourseAssignment.assignmentId == Assignment.id)\
-                                  .filter(CourseAssignment.courseId == course_id)\
-                                  .order_by(Submission.submissionDate).all()
-    for submisison in submissions:
-        submisison.assignmentName = Assignment.query.filter(Assignment.id == submisison.assignmentId).first().name
-    return submissions
-
-def getSubmissionResults(submissionId, submission):
-    submissionResults = SubmissionResult.query.filter(SubmissionResult.submissionId == submissionId).all()
-    if len(submissionResults) > 0:
-        assignmnet = Assignment.query.filter(Assignment.id == submission.assignmentId).first()
-        if assignmnet.isPublic:
-            return submissionResults
-        else:
-            # remove all results where the test cases were hidden
-            updatedSubmissionResults = []
-            for submissionResult in submissionResults:
-                testCase = TestCase.query.filter(TestCase.id == submissionResult.testCaseId).first()
-                if testCase.visible:
-                    updatedSubmissionResults.append(submissionResult)
-            return updatedSubmissionResults
-    return []
-
-def get_test_Cases(isOwner, assignmentId):
-    testCases = []
-    if isOwner:
-        testCases = TestCase.query.filter(TestCase.assignmentId == assignmentId).all()
-    else:
-        testCases = TestCase.query.filter(and_(TestCase.assignmentId == assignmentId, TestCase.visible == 1)).all()
-    return testCases
-
-def addTestCaseLog(filename_with_user_id, user, assignment_id, visible, maxScore):
-    test_case = TestCase(visible=visible, assignmentId=assignment_id, userId=user.id, submissionDate=func.now(), maxScore=maxScore)
-    db.session.add(test_case)
-    db.session.flush()  # To get test_case.id for the new entry
-    # testCase = TestCase(assignmentId = assignment_id, visible=visible, userId=user.id, fileName=filename_with_user_id, submissionDate=func.now())
-    # db.session.add(testCase)
-    # db.session.commit()
-    return True
-
-def togglevisiblity(test_case_id):
-    test_case = TestCase.query.filter_by(id=test_case_id).first()
-    if test_case:
-        test_case.visible = not test_case.visible
-        db.session.commit()
-
-def addTestCaseFileEntry(testcase_id, assignment_id, filename):
-    test_case_file = TestCaseFile(testCaseId=testcase_id, assignmentId=assignment_id, fileName=filename)
-    db.session.add(test_case_file)
-    db.session.commit()
-
-    return True
-
-def create_testcase(assignment_id, userId, visible, maxScore):
-    # Query the number of existing TestCase objects for this assignmentId
-    existing_test_cases_count = TestCase.query.filter_by(assignmentId=assignment_id).count()
-    
    
-    # Generate the new test case name
-    new_test_case_name = f"Test Case {existing_test_cases_count + 1}"
-    
-    # Create a new TestCase object
-    new_test_case = TestCase(
-        visible=visible,
-        assignmentId=assignment_id,
-        userId=userId,
-        submissionDate=func.now(),
-        maxScore=maxScore,
-        # Assuming you add a 'name' field to TestCase for storing "Test Case _"
-        name=new_test_case_name  
-    )
-    
-    # Add the new TestCase to the database session and commit it
-    db.session.add(new_test_case)
-    db.session.commit()
-    
-    return new_test_case
-
-
-def submit_to_moss(submission_directory, assignmentId):
-    userid = 732044316  # Your Moss user ID
-    m = mosspy.Moss(userid, "c")  # Specify "c" for C language
-
-    for root, dirs, files in os.walk(submission_directory):
-        for file in files:
-            if file.endswith(".c"):  # Ensure only C files are added
-                full_path = os.path.join(root, file)
-                m.addFile(full_path)
-    try:
-        url = m.send(lambda file_path, display_name: print('*', end='', flush=True))
-    except Exception as e:
-        print(f"Error sending files to Moss: {e}")
-    print("\nReport Url: " + url)
-
-    assignmnet = Assignment.query.filter(Assignment.id == assignmentId).first()
-    assignmnet.mossUrl = url
-    db.session.commit()
-
-    return url
-def compile_and_run_c_program(c_file_name, input_txt_name):
-    # Assuming current directory contains the files
-    current_dir = ""
-    c_file_path = os.path.join(current_dir, c_file_name)
-    input_txt_path = os.path.join(current_dir, input_txt_name)
-
-    # Compile the C program to an executable named 'student_output'
-    compile_command = ["gcc", c_file_path, "-o", "student_output"]
-    compile_result = subprocess.run(compile_command, capture_output=True)
-    if compile_result.returncode != 0:
-        # Handle compilation error properly
-        return None
-
-    # Run the compiled program with input from the txt file
-    with open(input_txt_path, 'r') as input_file:
-        run_command = ["./student_output"]
-        run_result = subprocess.run(run_command, stdin=input_file, text=True, capture_output=True)
-        if run_result.returncode == 0:
-            return run_result.stdout
-        else:
-            # Handle runtime error properly
-            return None
-
-def auto_grade(c_file_name, input_txt_name):
-    print("Grading submission...")
-    
-    # Run the grading logic
-    output = compile_and_run_c_program(c_file_name, input_txt_name)
-    
-    if output is not None:
-        print("Output of the student's program:")
-        print(output)
-    else:
-        print("Failed to compile or execute the student's program.")
-    
-    return 0
-
-def grade_submission(file_path, assignment_id, submission ):
-    #first test case:
-    print("1")
-    result = grade_submission2(file_path)
-    logGradingResult(result, "", 1, submission)
-
-    #second test case:
-    result = testCleanCompile(file_path)
-    logGradingResult(result, "", 2, submission)
-
-    #third test case:
-    notes = checkCode(file_path)
-    logGradingResult(result, notes, 3, submission)
-
-
-    return result
-
-def checkCode(c_file_path):
-    with open(c_file_path) as response:
-            answer = response.read()
-
-    notes = "\n"
-
-    notes = notes + 'Analysis of code:'
-
-
-    # check for usage of comments in student code
-    notes = notes +'\nComments:\n'
-    single_count = int(answer.count('//'))
-    multiple_count = int(answer.count('/*'))
-    sums = single_count + multiple_count
-    if sums >= 1:
-        notes = notes +str(sums) + ' comments used in the program.\n'
-    else:
-        notes = notes +'No comments used in the program.\n'
-
-
-
-   
-
-    #check for structures
-    notes = notes +'\nStructures:\n'
-    no_space = ''.join(answer.split())
-    structures_checked = ['for(', 'while(', 'if(', 'elseif(', 'else(', 'switch(']
-    for structure in structures_checked:
-        if int(no_space.count(structure) >= 1):
-            notes = notes + 'Structure: ' + structure + ' was found\n'
-
-
-
-    notes = notes +'\nVariables:\n'
-    x = 0
-    word = ''
-    while x < len(answer):
-        if answer[x].isspace():
-            variables = ['int', 'float', 'char', 'double', 'long']
-            for vari in variables:
-                if word == vari:
-                    while answer[x].isspace():
-                        x+=1
-                    word = ''
-                    while (not answer[x].isspace()) & (not answer[x] == ';'):
-                        if answer[x] == ',':
-                            s = vari + ' ' + word + ' found'
-                            if(len(word) > 6):
-                                if nonsense(word):
-                                    notes = notes +"\033[1;31m********************************************************************************\033[0m\n"
-                                    notes = notes +s + " - \033[1;31mNONSENSE. CHECK CODE.\033[0m\n"
-                                    notes = notes +"\033[1;31m********************************************************************************\033[0m\n"
-                                else:
-                                    notes = notes +s + " - \033[1;32mREAL WORD\033[0m\n"
-                            else:
-                                notes = notes +s + " - \033[1;34mCHECK CODE IF THIS IS NONSENSE, TOO SHORT FOR AUTODETECTOR\033[0m\n"
-                            x+=1
-                            while answer[x].isspace():
-                                x+=1
-                            word = ''
-                        word += answer[x]
-                        x+=1
-                    if word == 'main()':
-                        continue
-                    s = vari + ' ' + word + ' found'
-                    if(len(word) > 6):
-                        if nonsense(word):
-                            notes = notes +"\033[1;31m********************************************************************************\033[0m\n"
-                            notes = notes +s + " - \033[1;31mNONSENSE. CHECK CODE.\033[0m\n"
-                            notes = notes +"\033[1;31m********************************************************************************\033[0m\n"
-                        else:
-                            notes = notes +s + " - \033[1;32mREAL WORD\033[0m\n"
-                    else:
-                        notes = notes +s + " - \033[1;34mCHECK CODE IF THIS IS NONSENSE, TOO SHORT FOR AUTODETECTOR\033[0m\n"
-            word = ''
-        else:
-            word += answer[x]
-        x += 1
-
-
-
-    return notes
-
-def logGradingResult(result, notes, testCaseId, submission):
-    submission = Submission.query.filter(Submission.id == int(submission.id)).all()[0]
-    testCase = TestCase.query.filter(TestCase.id == testCaseId).all()[0]
-    gradedSubmission = SubmissionResult(assignmentId=submission.assignmentId, testCaseId=testCaseId, notes=notes, submissionId= submission.id, userId=submission.userId, gradeDate=func.now(), fileName=submission.fileName, score=result, maxScore = testCase.maxScore )
-    db.session.add(gradedSubmission)
-    db.session.commit()
-
-def testCleanCompile(c_file_path):
-    compile_process = subprocess.run(["gcc", c_file_path, "-o", "student_program"], capture_output=True, text=True)
-    if compile_process.returncode != 0:
-        # Handle compilation error properly, perhaps return a score of 0 or a specific error code
-        return 0
-    return 100
-
-
-def grade_submission2(c_file_path):
-    compile_process = subprocess.run(["gcc", c_file_path, "-o", "student_program"], capture_output=True, text=True)
-    if compile_process.returncode != 0:
-        # Handle compilation error properly, perhaps return a score of 0 or a specific error code
-        return 0
-
-    inputs = "5000.67 10000.89 25000.01\n2000.82 3000.01 500.33\n10"
-    expected_output_patterns = [
-        r"The cost of truck 1 after \d+ years is \$\d+\.?\d*",
-        # Add more patterns as needed
-    ]
-
-    run_process = subprocess.run(["./student_program"], input=inputs, capture_output=True, text=True)
-    if run_process.returncode != 0:
-        # Handle runtime error properly, perhaps return a score of 0 or a specific error code
-        return 0
-
-    actual_output = run_process.stdout
-
-    score = 100
-    penalty_per_error = 10
-
-    for pattern in expected_output_patterns:
-        if not re.search(pattern, actual_output, re.MULTILINE):
-            score -= penalty_per_error
-
-    # Return the final score as an integer
-    return score
         
-
-def assign_to_course(course_id, assignment_id, userId):
-    courseAssignment = CourseAssignment(courseId = course_id, assignmentId=assignment_id)
-    db.session.add(courseAssignment)
-    db.session.commit()
-    return True
-
-def find_user_assignments(userId):
-    assignments = Assignment.query.filter(Assignment.createdBy == userId).all()
-    return assignments
-
-def create_assignment(name, start_date, end_date, userId):
-    assignment = Assignment(name=name,startDate=start_date, endDate=end_date, createdBy = userId)
-    db.session.add(assignment)
-    db.session.commit()
-    return True
-
-def find_courses(user):
-    try:
-        # Explicitly specify the join condition and filter by userId
-        courses = db.session.query(Course).\
-            join(UserCourse, UserCourse.courseId == Course.id).\
-            filter(UserCourse.userId == user.id).all()
-        return courses
-    except Exception as e:
-        print(f"Error fetching courses for user {user.id}: {e}")
-        return []
-
-
 def pw_check(password):
     '''
     Ensure the password is valid
@@ -750,88 +318,6 @@ def length_check(str, min, max):
     if len(str) <= max and len(str) >= min:
         return True
     return False
-
-
-def update_user(curr_name, new_name, new_email, new_student_number, new_pw):
-    '''
-    R3-1, R3-4: Allow user to update username, password, email,
-    billing addr, and postal code.
-    Parameters:
-        curr_name   (String):     current username
-        new_name    (String):     updated username
-        new_email   (String):     updated email
-        new_pw      (String):     updated password
-    Returns:
-        True if the transaction is successful, False otherwise
-    '''
-    # If the current user exists
-    valid = User.query.filter_by(username=curr_name).all()
-    if len(valid) == 1:
-        # We check if the new information is of a valid format
-        if (
-            email_check(new_email) and
-            alphanumeric_check(new_name) and
-            length_check(new_name, 3, 19) and
-            pw_check(new_pw)
-        ):
-
-            # We then check if the new username and email are unique:
-            # If the user didn't update their existing names/passwords,
-            # the query will return 1, which is ok (it's their record),
-            # so ensure that the name and email have indeed been updated.
-            if (
-                ((len(User.query.filter_by(username=new_name).all()) > 0)
-                    and valid[0].username != new_name) or
-                ((len(User.query.filter_by(email=new_email).all()) > 0)
-                    and valid[0].email != new_email)
-            ):
-                return False
-            # If they're unique, update all the fields
-            else:
-                valid[0].username = new_name
-                valid[0].email = new_email
-                valid[0].student_number = new_student_number
-                valid[0].password = new_pw
-                db.session.commit()
-                return True
-        else:
-            # If any of the fields are not formatted properly, return False
-            return False
-    else:
-        # If the current user does not exist, return False right away
-        return False
-    
-def remove_testcase(testcase_id, assignment_id=None):
-    try:
-        # If assignment_id is provided, use it for extra validation
-        if assignment_id:
-            testcase = TestCase.query.filter_by(id=testcase_id, assignmentId=assignment_id).first()
-        else:
-            testcase = TestCase.query.filter_by(id=testcase_id).first()
-        
-        if testcase:
-            # find all TestCaseFiles
-            testCaseFiles = TestCaseFile.query.filter_by(testCaseId=testcase_id).delete()
-            db.session.delete(testcase)
-            db.session.commit()
-            return True
-        else:
-            return False  # Test case not found
-    except Exception as e:
-        print(f"Error removing test case: {e}")
-        return False
-
-def update_assignment_details(assignment_id, name, start_date, end_date, is_public):
-    assignment = Assignment.query.filter(Assignment.id == assignment_id).all()
-    if len(assignment) < 1:
-        return False
-    assignment = assignment[0]
-    assignment.name = name
-    assignment.start_date = start_date
-    assignment.end_date = end_date
-    assignment.isPublic = is_public
-    db.session.commit()
-    return True
 
 def not_empty(word):
     '''
